Implement themes for the control panel
Move some static content of the control panel to external modules.
At the moment there are two themes: select4cities and onesaitplatform.

Move resources to themes

Move some resources to the themes modules to allow diffent skins
in the control panel.

Configure control panel to use themes

Add configuration

Fix error in a directory name

The directory static/webjars was wrongly named.

Fix default profiles for maven

Default profiles have been removed. Now there are default values
for properties, and the build is based on those properties.
All the profiles have to be explicitaly used. The profiles override
the value of the properties when is required.