Improve the Encryption and Decryption of data
Implement AES encryption and decryption of data in ontologies.
A rest service is implemented in the IoTbroker to decrypt instances.
Add test to decrypt service in iotbroker
Add unit tests for the decrypt rest service in iotbroker.
The rest api of the iotbroker returns specific error codes when no
authorized user tries to decrypt data.


Former-commit-id: 4a93c2b4b07167218d621b7912c1e6d5b504c8ee